feat: sync

+++ project/SudMGPWrapper/Model/GameViewInfoModel.h
@@ -14,22 +14,24 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// 获取游戏View信息Model
 @interface ViewGameRect: NSObject
-@property (nonatomic, assign) CGFloat              left;
-@property (nonatomic, assign) CGFloat              top;
-@property (nonatomic, assign) CGFloat              right;
-@property (nonatomic, assign) CGFloat              bottom;
+@property (nonatomic, assign) NSInteger left;
+@property (nonatomic, assign) NSInteger top;
+@property (nonatomic, assign) NSInteger right;
+@property (nonatomic, assign) NSInteger bottom;
 @end
 
 @interface GameViewSize: NSObject
-@property (nonatomic, assign) CGFloat              width;
-@property (nonatomic, assign) CGFloat              height;
+@property (nonatomic, assign) NSInteger width;
+@property (nonatomic, assign) NSInteger height;
 @end
 
 @interface GameViewInfoModel: NSObject
-@property (nonatomic, strong) ViewGameRect          * view_game_rect;
-@property (nonatomic, strong) GameViewSize          * view_size;
-@property (nonatomic, assign) NSInteger             ret_code;
-@property (nonatomic, copy) NSString                * ret_msg;
+/// 游戏视图四周安全区 px
+@property (nonatomic, strong) ViewGameRect * view_game_rect;
+/// 游戏视图大小，设置与gameView保持一致大小值 px
+@property (nonatomic, strong) GameViewSize * view_size;
+@property (nonatomic, assign) NSInteger ret_code;
+@property (nonatomic, copy) NSString * ret_msg;
 /// 序列化成JSON格式字符串串
 - (nullable NSString *)toJSON;
 @end
